add TaskDoc creation from archive + context manager for loading archives
